generate_cut_list#11
